+++ extensions/prism-launcher/src/commands/join-server.tsx
@@ -29,7 +29,7 @@ import {
 import NoInstall from "./no-install";
 
 export default function JoinServer() {
-  const { data: isPrismInstalledData } = usePromise(isPrismLauncherInstalled, []);
+  const { data: isPrismInstalledData, isLoading: isPrismInstalledLoading } = usePromise(isPrismLauncherInstalled, []);
   const isPrismInstalled = isPrismInstalledData ?? false;
 
   const [instances, setInstances] = useState<Instance[]>();
@@ -172,14 +172,17 @@ export default function JoinServer() {
   return (
     <List
       searchBarPlaceholder={"Search instances..."}
-      {...(isPrismInstalled ? { isLoading: instances === undefined } : {})}
+      {...(isPrismInstalled ? { isLoading: instances === undefined } : { isLoading: isPrismInstalledLoading })}
     >
       <When condition={isPrismInstalled}>
         {instances?.map((instance, index) => (
           <List.Item
             key={`instance-${index}`}
             title={instance.name}
-            accessories={instance.favorite ? [{ icon: Icon.Star, tooltip: "Favorited" }] : []}
+            accessories={[
+              ...instance.accessories.map((acc) => ({ text: acc.text ?? undefined, icon: acc.icon ?? undefined })),
+              instance.favorite ? { icon: Icon.Star, tooltip: "Favorited" } : {},
+            ]}
             icon={{
               source: instance.icon ?? path.join(environment.assetsPath, "instance-icon.png"),
             }}