+++ extensions/prism-launcher/src/utils/preferences.ts
@@ -1,9 +1,4 @@
-import { type Application, getPreferenceValues } from "@raycast/api";
-
-type Preferences = {
-  installPath: Application;
-  instancesPath: string;
-};
+import { getPreferenceValues } from "@raycast/api";
 
 export function getPreferences<T extends keyof Preferences>(key: T): Preferences[T] {
   return getPreferenceValues<Preferences>()[key];