+++ extensions/prism-launcher/CHANGELOG.md
@@ -1,6 +1,12 @@
 # Prism Launcher Changelog
 
+## [New Addition] - {PR_MERGE_DATE}
+
+- Fixed a bug on windows where it wouldn't be able to get Prism Launcher
+- Added current versions for intances in launch instance and join server commands
+
 ## [New Addition] - 2025-11-24
+
 - Added support for custom instance folders
 - Added support for custom Prism Launcher installation paths
 - Refactored code to use usePromise hook for better async handling